拷贝dockercompose到back/version

+++ dchelp/image.py
@@ -5,8 +5,8 @@ import datetime
 img_back_path = './back/image/'
 temp_path = './temp/'
 def run_cmd(cmd):
-	print('执行shell:'+cmd)
-	p = subprocess.call(cmd, shell=True)
+  print('执行shell:'+cmd)
+  p = subprocess.call(cmd, shell=True)
 
 def read_images():
   dc_file = 'docker-compose.yml'
@@ -21,6 +21,8 @@ def read_images():
   return list(set(imgs))
 
 def do_image_pack():
+  p = "cp docker-compose.yml ./back/version/"
+  run_cmd(p)
   imgs = read_images()
   #打包镜像
   # docker save IMAGE > xxx.tar #  或者 docker save -o xxx.tar IMAGE
@@ -42,18 +44,13 @@ def do_image_pack():
 
 def do_image_unpack():
   imgs = read_images()
-  if not os.path.exists(temp_path):
-  run_cmd("mkdir "+temp_path)
+
+  unpack_list = {}
   for root, dirs, files in os.walk(img_back_path):  
     for file in files:
       if os.path.splitext(file)[1] == '.gz':
-        run_cmd("gzip -dc "+ root + file+" > "+ temp_path + os.path.splitext(file)[0])
-  unpack_list = {}
-  for root, dirs, files in os.walk(temp_path):  
-    for file in files:
-      if os.path.splitext(file)[1] == '.tar':
         file_name = '_'.join(os.path.splitext(file)[0].split('_')[:-1])
-        file_time = os.path.splitext(file)[0].split('_')[-1]
+        file_time = os.path.splitext(file)[0].split('_')[-1].split('.')[0]
         unpack_list[file_name] = file_time
   for image in imgs:
     pair_name = image.split("/")[-1]
@@ -64,15 +61,19 @@ def do_image_unpack():
         dt = datetime.datetime.strptime(res[:26],"%Y-%m-%dT%H:%M:%S.%f")
         time = dt.strftime("%Y%m%d%H%M%S")
         if long(unpack_list[pair_name]) > long(time):
-          p = "docker load < " + temp_path + file
+          if not os.path.exists(temp_path):
+            run_cmd("mkdir "+temp_path)
+          run_cmd("gzip -dc "+ img_back_path + pair_name + "_" + unpack_list[pair_name] + ".tar.gz" +" > "+ temp_path + pair_name + "_" + unpack_list[pair_name] + ".tar")
+          p = "docker load < " + temp_path + pair_name + "_" + unpack_list[pair_name] + ".tar.gz"
           run_cmd(p)
         else:
           print(image +"镜像已为最新!")
       else:
-        p = "docker load < " + temp_path + file
+        p = "docker load < " + temp_path + pair_name + "_" + unpack_list[pair_name] + ".tar.gz"
         run_cmd(p)
-  p = "rm -rf "+temp_path
-  run_cmd(p)
+  if os.path.exists(temp_path):
+    p = "rm -rf "+temp_path
+    run_cmd(p)
 
 def do_image_clear():
   """对进行进行清理
